added the ref to PA article and added a few more paragraphs
Jacob and Flow, we are getting pretty close to the page limit, but man
this is short. I don't think we are at 1000 words yet.  But with six or
so more paragraphs, a figure and a table plus a hand flu of references
to be added, we are almost done.