Allowing matrix of weights, making convergence test
Now, user is allowed to put in a matrix of weights. I’ve re-wrote the
checks at the top to handle both matrices and vectors of weights. I
also wrote code to calibrate a separate model for each set of weights.
Still need to calculate MAD of posterior weights, throw error if too
large.